pyproject.toml
[tool.poetry]
name = "aws-cloudwatch-metrics"
version = "0.1.0"
description = "Simple script to demo aws cloudwatch metrics"
authors = ["KeaganGilmore <keagangilmore@gmail.com>"]
license = "GPL-3.0 license"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
boto3 = "^1.35.76"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
